Added code sync feature to New UI
