+++ client/package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-
     "@clerk/nextjs": "^6.22.0",
     "@monaco-editor/react": "^4.7.0",
     "@uiw/react-codemirror": "^4.23.13",
@@ -22,7 +21,6 @@
     "react-hot-toast": "^2.5.2",
     "socket.io-client": "^4.8.1",
     "zustand": "^5.0.5"
-
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
