+++ server/src/app.ts
@@ -4,6 +4,7 @@ import { Server } from "socket.io";
 import cors from "cors";
 import { Webhooks } from "./utils/WebHooks";
 import bodyParser from "body-parser";
+// import { codeExecution } from "./models/Connection.model";
 
 //import routers
 import healthCheckRouter from "./routers/healthCheck.routers";
@@ -26,10 +27,45 @@ app.use(express.static("public"));
 // routes
 app.use("/api/health", healthCheckRouter);
 
+io.on("connection", (socket) => {
+  console.log("a user connected on socket", socket.data);
+
+  socket.on("join-room", async ({ roomId }) => {
+    socket.join(roomId);
+    console.log(`User ${socket.id} joined room ${roomId}`);
+    // const currentCode = await codeExecution.findOne({ roomId });
+    // if (currentCode) {
+    //   socket.emit("receive-changes", currentCode.currentCodeContent);
+    // }
+  });
+
+  socket.on("code-change", async ({ roomId, code }) => {
+    socket.to(roomId).emit("receive-changes", code);
+    // const currentCode = await codeExecution.findOne({ roomId });
+    // if (currentCode) {
+    //   currentCode.currentCodeContent = code;
+    //   await currentCode.save();
+    // } else {
+    //   const newCode = new codeExecution({
+    //     roomId: roomId,
+    //     currentCodeContent: code,
+    //   });
+    //   await newCode.save();
+    // }
+  });
+
+  socket.on("disconnect", () => {
+    console.log("user disconnected from socket");
+  });
+});
+
 app.post(
   "/api/webhook",
   bodyParser.raw({ type: "application/json" }),
-  Webhooks
+  (req, res) => {
+    console.log("webhook called");
+    Webhooks(req, res);
+  }
 );
 
 export { httpServer, io };
